requirements-oscar2.txt
alabaster==0.7.10
Babel==2.5.1
behave==1.2.5
behave-django==0.5.0
Django==1.11.3
django-appconf==1.0.2
django-bootstrap3==9.0.0
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.8.1
django-hamlpy==1.1.1
django-pdb==0.6.1
docutils==0.14
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.6
future==0.16.0
imagesize==0.7.1
ipaddress==1.0.18
ipdb==0.8.1
ipython==3.2.0
Jinja2==2.9.6
Markdown==2.6.2
MarkupSafe==1.0
mechanize==0.2.5
numpy==1.13.3
pandas==0.20.3
parse==1.8.2
parse-type==0.4.2
Pygments==2.0.2
PyHamcrest==1.9.0
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.11
rcssmin==1.0.6
regex==2017.9.23
requests==2.10.0
rjsmin==1.0.12
ruamel.ordereddict==0.4.9
ruamel.yaml==0.12.13
six==1.9.0
snowballstemmer==1.2.1
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.3
text-unidecode==1.0
typing==3.6.2
Werkzeug==0.10.4
xlrd==1.1.0
yamlordereddictloader==0.1.1
#gnureadline==6.3.3; sys_platform != 'win32'
#pyreadline==2.1; sys_platform == 'win32'
psycopg2==2.7.3.2
selenium==3.6.0
coveralls==1.2.0
coverage==4.0.3
